+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Changed
 
 - changed some function names to conform to PEP8 standards
+- import entire constants file to `main.py` instead of every constant needed
 
 ### Removed